+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - bootstrap: enable default S3 server-side encryption
+- output `aws-enterprise-vpc_version`
 
 ### Changed
 - vpc: extend workaround to accommodate a third AZ in transit_gateway_attachment_subnets